HLE_OS: Remove HLE_VPrintf and HLE_GeneralDebugPrintWithInt
These were originally added in 57e3fbe and bd27982, but were later
removed from the table in a merge conflict patch, 934e150. Matt
says that it was intentional that they were removed from the table
in that commit, since the code was really only intended for a few
games while he was debugging. Delete the vestigial functions.

If somebody wants these back, they can revert this commit and update
the table.